+++ src/containerapp/azext_containerapp/_params.py
@@ -48,7 +48,7 @@ def load_arguments(self, _):
         c.argument('revision', help="The name of the container app revision. Defaults to the latest revision.")
         c.argument('name', name_type, id_part=None, help="The name of the Containerapp.")
         c.argument('resource_group_name', arg_type=resource_group_name_type, id_part=None)
-        c.argument('kind', options_list=["--kind", "-k"], help="Type of logs to stream", arg_type=get_enum_type([LOG_TYPE_CONSOLE, LOG_TYPE_SYSTEM]), default=LOG_TYPE_CONSOLE)
+        c.argument('kind', options_list=["--type", "-t"], help="Type of logs to stream", arg_type=get_enum_type([LOG_TYPE_CONSOLE, LOG_TYPE_SYSTEM]), default=LOG_TYPE_CONSOLE)
 
     with self.argument_context('containerapp env logs show') as c:
         c.argument('follow', help="Print logs in real time if present.", arg_type=get_three_state_flag())