+++ src/containerapp/HISTORY.rst
@@ -7,7 +7,7 @@ Release History
 ++++++
 * Add 'az containerapp env update' to update managed environment properties
 * Add custom domains support to 'az containerapp env create' and 'az containerapp env update'
-* 'az containerapp logs show': add new parameter "--kind" to allow showing system logs
+* 'az containerapp logs show': add new parameter "--type" to allow showing system logs
 * Show system environment logs with new command 'az containerapp env logs show'
 * Add tcp support for ingress transport and scale rules
 * `az containerapp up/github-action add`: Retrieve workflow file name from github actions API